doing HW2

+++ ent/schema/expense.go
@@ -21,7 +21,7 @@ func (Expense) Mixin() []ent.Mixin {
 // Fields of the Expense.
 func (Expense) Fields() []ent.Field {
 	return []ent.Field{
-		field.Int("amount").
+		field.Float("amount").
 			Positive(),
 		field.String("category").Default("UNCATEGORIZED"),
 	}
